Create Ixantripx-mix-love-advanture
The President of the State of Israel (Hebrew: נָשִׂיא מְדִינַת יִשְׂרָאֵל‬, Nasi Medinat Yisra'el, or Hebrew: נָשִׂיא הַמְדִינָה‬, Nasi HaMedina, literally President of the State) is the head of state of Israel. The position is largely a ceremonial figurehead role,[1] with executive power effectively being exercised by the Prime Minister. The current president is Reuven Rivlin, who took office on 24 July 2014. Presidents are elected by the Knesset for a seven-year term and are limited to a single term.